GLTF系モジュールを分割した
