+++ src/js/resource/gltf/ids.ts
@@ -1,0 +1,4 @@
+/** IDリスト */
+export const GLTF_IDS = {
+  SHOPPING_STREET: "SHOPPING_STREET",
+};
