+++ src/js/resource/loading/full-resource-configs.ts
@@ -1,6 +1,6 @@
 import { CANVAS_IMAGE_CONFIGS } from "../canvas-image";
 import { CUBE_TEXTURE_CONFIGS } from "../cube-texture";
-import { GLTF_CONFIGS } from "../gltf";
+import { GLTF_CONFIGS } from "../gltf/configs";
 import { SOUND_CONFIGS } from "../sound/configs";
 import {
   DEVELOPING_TEXTURE_CONFIGS,
