+++ src/js/resource/loading/resource-loading.ts
@@ -5,8 +5,8 @@ import type { CanvasImageConfig, CanvasImageResource } from "../canvas-image";
 import { loadCanvasImage } from "../canvas-image";
 import type { CubeTextureConfig, CubeTextureResource } from "../cube-texture";
 import { loadCubeTexture } from "../cube-texture";
-import type { GlTFConfig, GlTFResource } from "../gltf";
-import { loadGlTF } from "../gltf";
+import { loadGlTF } from "../gltf/load-gltf";
+import { GlTFConfig, GlTFResource } from "../gltf/resource";
 import { getAllPaths } from "../path/get-all-paths";
 import { preLoadImage } from "../path/pre-load-image";
 import { PathConfig } from "../path/resource";
